Make rootfs mounting more consistent and robust
* Remounting the rootfs as read-write was broken and no one noticed it
yet. There was a stray $ in the path we were checking for the
.writable_image file.
* The Debian initramfs sets environment variables based on cmdline
parameters. Some android devices have the 'syspart=' parameter passed by
the bootloader. Because of this, I've renamed $syspart to $_syspart.
* Remove the remount of '/halium-system' as read-only when doing an
Android boot. If the rootfs is already read-write, it is likely that we
have a file handle open on it by this point and won't be able to remount
it read-only anyway. If the remount operation doesn't check for open
file handles, we could potentially corrupt the Halium rootfs by doing
this.