halium: Support system installation on A/B
On devices with A/B partitioning scheme it can be required,
due to the system partname not pointing to the suffixed partition,
to manually point the initramfs mounting section to mount the
suffixed partlabel.

To make use of this, the systempart kernel argument can be set
with a slotselect prefix, ie 'systempart=slotselect:system',
which will then look up the right 'system' partlabel manually.